Implement basic DrawTriplanar pass for terrain
Modified the DrawShaded shader to do simple triplanar rendering.
Currently it's a little hacky because we put the material in a
TriplanarMaterial struct to avoid the regular DrawShaded pass from
seeing it. Furthermore we use the albedo and emission textures for
the horizontal and vertial texture respectively. Really our triplanar
material should be a combination of 3/4 regular materials, but
re-purposing the existing material fields for our purpose is much
simpler for now.

Allow passing 3 materials to DrawTriplanar

Currently using emission textures for various other things like passing
in a noise texture, but should probably figure a better way of doing
this

Whiteout blend for triplanar normal maps